requirements.txt
torch
torchvision
torchaudio
numpy
pyautogui
rich
pyyaml
pillow
keyboard
pyinstaller
PyQt6